added helmet & cyclic dynamo db, rework auth to rbac

+++ index.js
@@ -6,16 +6,24 @@ const app = express()
 const AWS = require("aws-sdk")
 const s3 = new AWS.S3()
 
+// cyclic db
+const db = require('cyclic-dynamodb')
+const users = db.collection('users')
+
 // misc
 const cors = require('cors')
 
-//auth
+// security
 const jwt = require('jsonwebtoken')
+const helmet = require('helmet')
 
+app.use(helmet())
 app.use(express.json())
+app.use(express.urlencoded({ extended: true }))
 app.use(cors())
 
 // GET https://some-app.cyclic.app/files?name=
+// get specific ".json" with the filename
 app.get('/api/files', async (req, res) => {
   const filename = req.query.name + '.json'
 
@@ -49,18 +57,46 @@ app.get('/api/listJson', async (req, res) => {
 
 })
 
+app.get('/api/listUser', async (req, res) => {
+  const userObjects = await users.list().promise()
+
+  res.json(userObjects).end()
+})
+
+app.post('/api/signup', async (req, res) => {
+  console.log(req.body)
+
+  const email = req.body.email
+  const password = req.body.password
+  const role = 'dev'
+  const uid = 'uid_' + Math.random().toString().slice(2)
+  const uProps = {
+    uid,
+    email,
+    password,
+    role
+  }
+
+  const user = await users.set(email, uProps, { $index: ['uid'] })
+
+  res.json(user.props).end()
+})
+
 // PROTECTED
 const accessTokenSecret = process.env.SECRET_TOKEN
 app.post('/api/signin', (req, res) => {
-  const { password } = req.body;
-  if (password === process.env.PASSWORD) {
+  const { username, password } = req.body;
+
+  const user = users.find(u => { return u.username === username && u.password === password })
+
+  if (user) {
     // Generate an access token
-    const accessToken = jwt.sign({ password }, accessTokenSecret, { expiresIn: '20m' })
+    const accessToken = jwt.sign({ username: user.username, role: user.role }, accessTokenSecret)
 
     res.json({ accessToken })
   }
   else
-    res.sendStatus(401).end()
+    res.send('Incorrect username or password')
 
 })
 
@@ -71,13 +107,13 @@ const authenticateJWT = (req, res, next) => {
   if (authHeader) {
     const token = authHeader.split(' ')[1]
 
-    jwt.verify(token, accessTokenSecret, (err, password) => {
+    jwt.verify(token, accessTokenSecret, (err, user) => {
       if (err)
         return res.sendStatus(403)
 
-      req.password = password
+      req.user = user
       next()
-    });
+    })
   }
   else
     res.sendStatus(401).end()
@@ -85,6 +121,11 @@ const authenticateJWT = (req, res, next) => {
 
 // PUT https://some-app.cyclic.app/api/files?name=
 app.put('/api/files', authenticateJWT, async (req, res) => {
+  const { role } = req.user
+
+  if (role !== admin)
+    return res.sendStatus(403)
+
   const filename = req.query.name + '.json'
 
   await s3.putObject({
@@ -99,6 +140,11 @@ app.put('/api/files', authenticateJWT, async (req, res) => {
 
 // DELETE https://some-app.cyclic.app/api/files?name=
 app.delete('/api/files', authenticateJWT, async (req, res) => {
+  const { role } = req.user
+
+  if (role !== admin)
+    return res.sendStatus(403)
+
   const filename = req.query.name + '.json'
 
   await s3.deleteObject({
@@ -113,13 +159,12 @@ app.delete('/api/files', authenticateJWT, async (req, res) => {
 // /////////////////////////////////////////////////////////////////////////////
 // Catch all handler for all other request.
 app.use('*', (req, res) => {
-  res.send('No endpoint listening here')
-  res.sendStatus(404).end()
+  res.json({ msg: 'no route handler found', path: req.path, method: req.method }).end()
 })
 
 // /////////////////////////////////////////////////////////////////////////////
 // Start the server
 const port = process.env.PORT || 3000
 app.listen(port, () => {
-  console.log(`index.js listening at http://localhost:${port}`)
+  console.log(`index.js listening on ${port}`)
 })
